resolve refs and refactor code
