+++ pkg/controller/direct/monitoring/refs.go
@@ -19,6 +19,8 @@ import (
 	"fmt"
 	"strings"
 
+	"github.com/GoogleCloudPlatform/k8s-config-connector/pkg/controller/direct/resolverefs"
+
 	krm "github.com/GoogleCloudPlatform/k8s-config-connector/apis/monitoring/v1beta1"
 	refs "github.com/GoogleCloudPlatform/k8s-config-connector/apis/refs/v1beta1"
 	"github.com/GoogleCloudPlatform/k8s-config-connector/pkg/clients/generated/apis/k8s/v1alpha1"
@@ -54,7 +56,7 @@ func normalizeResourceName(ctx context.Context, reader client.Reader, src client
 
 	switch ref.Kind {
 	case "Project":
-		project, err := refs.ResolveProject(ctx, reader, src, &refs.ProjectRef{
+		project, err := resolverefs.ResolveProjectRef(ctx, reader, src, &refs.ProjectRef{
 			Name:      ref.Name,
 			Namespace: ref.Namespace,
 			External:  ref.External,
@@ -88,7 +90,7 @@ func normalizeResourceName(ctx context.Context, reader client.Reader, src client
 	return ref, nil
 }
 
-func normalizeMonitoringAlertPolicyRef(ctx context.Context, reader client.Reader, src client.Object, project refs.Project, ref *refs.MonitoringAlertPolicyRef) (*refs.MonitoringAlertPolicyRef, error) {
+func normalizeMonitoringAlertPolicyRef(ctx context.Context, reader client.Reader, src client.Object, project resolverefs.Project, ref *refs.MonitoringAlertPolicyRef) (*refs.MonitoringAlertPolicyRef, error) {
 	if ref == nil {
 		return nil, nil
 	}
@@ -138,12 +140,12 @@ func normalizeMonitoringAlertPolicyRef(ctx context.Context, reader client.Reader
 		return nil, fmt.Errorf("error reading referenced MonitoringAlertPolicy %v: %w", key, err)
 	}
 
-	alertPolicyResourceID, err := refs.GetResourceID(alertPolicy)
+	alertPolicyResourceID, err := resolverefs.GetResourceID(alertPolicy)
 	if err != nil {
 		return nil, err
 	}
 
-	alertPolicyProjectID, err := refs.GetProjectID(ctx, reader, alertPolicy)
+	alertPolicyProjectID, err := resolverefs.GetProjectID(ctx, reader, alertPolicy)
 	if err != nil {
 		return nil, err
 	}
@@ -160,7 +162,7 @@ func normalizeProjectRef(ctx context.Context, reader client.Reader, src client.O
 		return nil, nil
 	}
 
-	project, err := refs.ResolveProject(ctx, reader, src, ref)
+	project, err := resolverefs.ResolveProjectRef(ctx, reader, src, ref)
 	if err != nil {
 		return nil, err
 	}
@@ -174,7 +176,7 @@ type refNormalizer struct {
 	ctx     context.Context
 	kube    client.Reader
 	src     client.Object
-	project refs.Project
+	project resolverefs.Project
 }
 
 func (r *refNormalizer) VisitField(path string, v any) error {
